Fix id field being incorrectly added to some queries
The id field was being included in the query even when not required,
i.e. even when there is a single query step and the data is not going to
be merged with another step.
And because in that case the result just gets forwarded as is,
the field was not getting removed during formatting.

This fixes query planning to not add the id field when the data won't be
merged, or in other words:
- The step is not a child step; AND
- The step doesn't have any child step