Unmarshal only what's necessary when merging results
When dealing with fragment spreads we usually want to pass the response
from the client as is, because we don't necessarily know what type was
matched. In order to do that we need to not unmarshal the response more
than necessary.

So when merging results maps, unmarshal into a
`map[string]json.RawMessage`  instead of `map[string]interface{}`, this
way we recursively unmarshal only what's strictly necessary to do the
merging.